+++ CHANGELOG.md
@@ -8,7 +8,7 @@ All notable changes to Rainbow-CLI will be documented in this file.
 
 ## [1.81.0] - 2020-XX-XX
 - New command to get application analytics 'rbw application analytics'
-- New commands to get bot services 'rbw bot services [companyId]', 'rbw bot service <serviceId> [companyId]'
+- Bug fix in rbw dashboard command
 
 ## [1.71.0] - 2020-04-14
 - New option to create a guest user 'rbw create user --guest'